Rename from reactivesocket (old) to rsocket (new).
