+++ .gitignore
@@ -90,3 +90,4 @@ ENV/
 
 # PyCharm
 *.iml
+.idea/
