+++ docs/index.rst
@@ -1,17 +1,17 @@
-.. reactivesocket documentation master file, created by
+.. RSocket documentation master file, created by
    sphinx-quickstart on Thu Feb  9 05:26:44 2017.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to reactivesocket's documentation!
+Welcome to RSocket's documentation!
 ==========================================
 
 .. toctree::
    :maxdepth: 2
    :caption: Contents:
 
-.. automodule:: reactivesocket
-.. autoclass:: reactivesocket.RequestHandler
+.. automodule:: rsocket
+.. autoclass:: rsocket.RequestHandler
    :members:
 
 .. automodule:: reactivestreams
