--- reactivesocket/__init__.py
@@ -1,11 +1,0 @@
-"""
-ReactiveSocket
-~~~~~~~~~~~~~~
-
-A ReactiveSocket implementation in Python.
-"""
-from reactivesocket.reactivesocket import RequestHandler, BaseRequestHandler
-from reactivesocket.reactivesocket import ReactiveSocket
-from reactivesocket.handlers import RequestResponseRequester, StreamHandler
-from reactivesocket.payload import Payload
-from reactivestreams import Publisher, Subscriber, Subscription
