+++ tests/rsocket/conftest.py
@@ -1,20 +1,20 @@
 import asyncio
 import pytest
 
-from reactivesocket import ReactiveSocket
+from rsocket import RSocket
 
 
 @pytest.fixture
 def pipe(unused_tcp_port, event_loop):
     def session(reader, writer):
         nonlocal server
-        server = ReactiveSocket(reader, writer)
+        server = RSocket(reader, writer)
 
     async def start():
         nonlocal service, client
         service = await asyncio.start_server(session, host, port)
         connection = await asyncio.open_connection(host, port)
-        client = ReactiveSocket(*connection, server=False)
+        client = RSocket(*connection, server=False)
 
     async def finish():
         service.close()
@@ -28,5 +28,3 @@ def pipe(unused_tcp_port, event_loop):
     event_loop.run_until_complete(start())
     yield server, client
     event_loop.run_until_complete(finish())
-
-
